feat(latex): compose documents by means of `String`s
A base template may not be the right way to deal with dynamic contents. There may be sections we don't want, there may be entries which we don't want, so having a fixed template may not allow this flexibility.

Since the template rendering returns a `String` we can concatenate them as we need, as well as filter all the data as we need.